Move crashNow helpers to NewRelic+Helpers.swift
Addresses review feedback: keeps NewRelic.swift from opening with the
crash-testing helper, following this codebase's existing +Extension.swift
convention (e.g. UIColor+HexString.swift). No behavior change — verified
via full rebuild and the NewRelicTests/NewRelicAPITest suites (35/35).
